+++ src/View/view_info_updater.h
@@ -47,10 +47,10 @@ class ViewInfoUpdater {
   QLabel* start_label_ = nullptr;
   QLayout* layout_ = nullptr;
 
-  uint32_t laps_amount_ = 0;
-  uint32_t players_amount_ = 0;
-  int32_t seconds_before_start_ = 5;
-  int32_t network_id_ = 0;
+  size_t laps_amount_ = 0;
+  size_t players_amount_ = 0;
+  size_t seconds_before_start_ = 5;
+  size_t network_id_ = 0;
   bool is_game_started_ = false;
   CarsData cars_data_;
 };
